% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale.R
\name{colorscaleThreeJS}
\alias{colorscale}
\alias{colorscaleThreeJS}
\alias{is.colorscaleThreeJS}
\alias{toJSONList.colorscaleThreeJS}
\title{S3 class for threeJS LUT colorscale}
\usage{
colorscaleThreeJS(colorMap = c("rainbow", "cooltoworm", "blackbody",
  "greyscale"), numcolors = 512, max = 1, min = 0, legend.on = F,
  legend.layout = c("vertical", "horizontal"), legend.position = c(0.9,
  0.8), legend.dimensions = c(0.05, 0.3), legend.labels.on = F,
  legend.labels.size = 24, legend.labels.font = "Arial",
  legend.labels.title = "", legend.labels.titleColor = "white",
  legend.labels.um = NULL, legend.labels.nticks = 8,
  legend.labels.tickColor = "white", legend.labels.labelColor = "white",
  legend.labels.decimal = 2, legend.labels.notation = c("standard",
  "scientific"))

is.colorscaleThreeJS(obj)

\method{toJSONList}{colorscaleThreeJS}(obj)
}
\arguments{
\item{colorMap}{Color scale mapping}

\item{numcolors}{Number of different colors}

\item{max}{Maximum value in the scale}

\item{min}{Minimim value in the scale}

\item{legend.on}{If \code{TRUE} a colorbar legend is added to the scene}

\item{legend.layout}{Either vertical or horizontal}

\item{legend.position}{Legend position in the range [0,1] x [0,1] in top-left to bottom-right orientation}

\item{legend.dimensions}{Legend size in the x,y dimensions from 0 to 1}

\item{legend.labels.on}{Add labels to the ticks in the legend}
}
\description{
An R-side object of a JS LUT colorscale. It includes a set of predefined color maps,
and a huge ammount of parameters. Color legend can be added.
}
\examples{
\dontrun{
#colorscaleThreeJS(
#max = 1, min = -1, legend.on = T, legend.labels.on = T,
#legend.labels.title = 'EXAMPLE')
runShinyExample("parametric")
}

}

