% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helperThreeJS}
\alias{axisHelper}
\alias{boundingBox}
\alias{gridHelper}
\alias{helper}
\alias{helperThreeJS}
\alias{is.helperThreeJS}
\alias{toJSONList.helperThreeJS}
\title{S3 class for threeJS helpers}
\usage{
helperThreeJS(builder, type = c("global", "mesh"), color = "#FFFFFF",
  size = 1)

is.helperThreeJS(obj)

\method{toJSONList}{helperThreeJS}(obj)

boundingBox(color = "#FFFF00", diagonals = F)

axisHelper(size = 25, center = c(0, 0, 0))

gridHelper(size = 10, step = 1, gridColor = "#FFFFFF",
  midColor = "#FFFF00", center = c(0, 0, 0), rotation = c(0, 0, 0))
}
\arguments{
\item{builder}{Javascript function that builds the helper}

\item{type}{Either global or mesh}

\item{color}{Helper color in RGB format}

\item{size}{Helper size}

\item{diagonals}{If true, diagonals on each face of the box are included}

\item{center}{3D vector that locates the center of the helper}

\item{step}{Numeric value. Size of the step between two lines}

\item{gridColor}{Color of the lines of the grid.}

\item{midColor}{Color of the centerline.}

\item{size}{Numeric value. axis length}

\item{size}{Numeric value. Grid size}

\item{Rotation}{3D vector that specifies the x,y,z rotations}
}
\description{
A generic threeJS helper definition. Helpers can be "global" or "mesh-specifc"
}
\details{
\code{boundingBox} Creates a bounding box instance. This helper is mesh(object)-specific

\code{axisHelper} Creates a Axis helper instance. This is a global helper (not associated with a specific object)

\code{gridHelper} Creates a Grid helper. This helper is not linked to a specific object
}
\examples{
 obj.helper <- boundingBox()
 obj <- objThreeJS(sphereGeometry(1, position = c(1,1,1)),
     lambertMaterial(color = "#FFFFFF"), helpers = obj.helper)
 grid <- gridHelper(center = c(0,-1,0))
 axis <- axisHelper(size = 10)

}

