% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{JSONThreeArgList}
\alias{JSONThreeArgList}
\title{JSONThreeArgList}
\usage{
JSONThreeArgList(obj)
}
\arguments{
\item{obj}{The original object}
}
\description{
Creates the argument vector for a threeJS builder function
}
\details{
Creates an unnamed list with the arguments needed for the builder func
}

