% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{lightThreeJS}
\alias{ambientLight}
\alias{directionalLight}
\alias{light}
\alias{lightThreeJS}
\title{ThreeJS Lights}
\usage{
lightThreeJS(builder, color = "#FFFFFF")

ambientLight(color = "#ffffff")

directionalLight(color = "#666666", intensity = 1, from = c(10, 0, 10))
}
\arguments{
\item{builder}{Javascript function that builds the lights}

\item{color}{Light color in RGB format}

\item{intensity}{0-1 intensity value}

\item{from}{3D numeric vector (x,y,z) (light direction)}

\item{color}{Material color (either a RGB string or a valid name)}
}
\description{
Light implementations such as Ambient or Directional lighting.
}
\details{
\code{ambientLight} creates a  ambient light instance, that is, global uniform lighting

\code{directionalLight} Creates a directional light instance with color, intensity and position parameters
}
\examples{
light <-
list(
 directionalLight(
   "#FFFFFF",intensity = 0.5,from = c(1,1,0)
 ),
 directionalLight(
   "#FFFFFF",intensity = 0.5,from = c(0,0,1)
   #'  ),
 ambientLight("#333333")
)

}

