% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{objThreeJS}
\alias{objThreeJS}
\alias{object}
\title{ThreeJS 3DObject instance}
\usage{
objThreeJS(geometries, material, name = NULL, helpers = NULL, ...)
}
\arguments{
\item{geometries}{List of geoThreeJS to be included in this object (see \link{geometry}).}

\item{material}{Object material instance (see \link{material}).}

\item{name}{Optional name of the object. Doesnt need to be unique}

\item{helpers}{List of object helpers (see \link{helper}).}

\item{\dots}{Children 3D objects}
}
\description{
A threeJS 3D Object definition. An object is defined by a geometry (see \link{geometry}) and
a surface material (see \link{material}). Opitionally you can add a object-specific helper
 (see \link{helper}) that will affect this object and all its descendants.
}
\examples{
objThreeJS( sphereGeometry(1), lambertMaterial(color = "red"))
}

