% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejswrapper.R
\docType{package}
\name{threejswrapper-package}
\alias{threejswrapper}
\alias{threejswrapper-package}
\title{Interactive 3D plots for R using three.js and htmlwidgets.}
\description{
Interactive 3D plots for R using three.js and htmlwidgets.
}
\examples{
\dontrun{
 runShinyExample("spheres")
 runShinyExample("cylinders")
 runShinyExample("torus")
 runShinyExample("mesh")
 runShinyExample("parametric")
}
}
\references{
\url{http://threejs.org}
}
\seealso{
\link{scene} \link{threejswrapper-shiny}
}

