% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{toJSONList}
\alias{toJSONList}
\title{toJSONList}
\usage{
toJSONList(obj)
}
\arguments{
\item{obj}{The original object}
}
\description{
Transform the object to a list
}
\details{
Transform the given object to a list suitable for toJSON conversion
by any conventional package such as RJSONIO or jsonlite
}

